+++ CHANGELOG.md
@@ -4,6 +4,12 @@ fractal-fork version history
 unreleased
 ----------
 
+* tbd
+
+
+0.0.3
+-----
+
 * [patch] upgraded following dependencies: `liftoff`, `globby`
 * [patch] removed `istextorbinary` to check if a file is binary and always parse with `utf-8`
 * [patch] removed `columnify` since `console.columns` method is never used